backend: auth routes and memory routes with auth middleware (actually a preHandler hook)

+++ server/src/routes/memories.ts
@@ -3,8 +3,16 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
+  // fastify.io/docs/latest/Reference/Lifecycle/ explains the http req lifecycle
+  // handlers are the functions defined in routes defined below
+  // so we want to add logic before every handler func
+  app.addHook('preHandler', async (req, res) => {
+    await req.jwtVerify()
+  })
+
   app.get('/memories', async (req, res) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: req.user.sub },
       orderBy: { createdAt: 'asc' },
     })
 
@@ -18,18 +26,25 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   app.get('/memories/:memoryId', async (req, res) => {
-    // define memoryId format
+    // define memoryId format && validate if memoryId is a valid uuid string
     const paramsSchema = z.object({ memoryId: z.string().uuid() })
-    // validate if memoryId is a valid uuid string
     const { memoryId } = paramsSchema.parse(req.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id: memoryId },
     })
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return res.status(401).send({
+        message:
+          "We still do not have the technology to provide access to somebody else's memory",
+      })
+    }
     return memory
   })
 
   app.post('/memories', async (req, res) => {
+    // define & validate memory req.body
     const bodySchema = z.object({
       content: z.string(),
       // coerce will convert whatever comes here to boolean --> Boolean(val)
@@ -43,13 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         isPublic,
         coverUrl,
-        userId: '81666a73-c530-4cf4-86f6-6a30eafcc29b',
+        userId: req.user.sub,
       },
     })
     return memory
   })
 
   app.put('/memories/:memoryId', async (req, res) => {
+    // define & validate memory req.params
     const paramsSchema = z.object({ memoryId: z.string().uuid() })
     const { memoryId } = paramsSchema.parse(req.params)
 
@@ -60,6 +76,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
       coverUrl: z.string(),
     })
     const { content, isPublic, coverUrl } = bodySchema.parse(req.body)
+
+    const memoryFound = await prisma.memory.findUniqueOrThrow({
+      where: { id: memoryId },
+    })
+    if (memoryFound.userId !== req.user.sub) {
+      return res.status(401).send({
+        message:
+          "We still do not have the technology to let you edit somebody else's memory",
+      })
+    }
 
     const memoryEdited = await prisma.memory.update({
       where: { id: memoryId },
@@ -73,8 +99,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
   app.delete('/memories/:memoryId', async (req, res) => {
+    // define & validate memory req.params
     const paramsSchema = z.object({ memoryId: z.string().uuid() })
     const { memoryId } = paramsSchema.parse(req.params)
+
+    const memoryFound = await prisma.memory.findUniqueOrThrow({
+      where: { id: memoryId },
+    })
+    if (memoryFound.userId !== req.user.sub) {
+      return res.status(401).send({
+        message:
+          "We still do not have the technology to let you edit somebody else's memory",
+      })
+    }
 
     const memoryDeleted = await prisma.memory.delete({
       where: { id: memoryId },
